Minor code improvements

+++ check_pa/modules/throughput.py
@@ -8,8 +8,10 @@ import time
 import nagiosplugin as np
 
 from check_pa.xml_reader import XMLReader, Finder
+from check_pa.utils import *
 
 _log = logging.getLogger('nagiosplugin')
+statefile = os.path.join(tempfile.gettempdir(), 'throughput')
 
 
 def get_time():
@@ -23,8 +25,9 @@ def get_time():
 
 def create_check(args):
     """
+    Creates and configures a check for the throughput command.
 
-    :return:
+    :return: the throughput check.
     """
     interfaces = str(args.interface).split(",")
     check = np.Check()
@@ -37,16 +40,19 @@ def create_check(args):
     check.add(NetworkSummary())
     return check
 
-def reset(): # pragma: no cover
+
+def reset():  # pragma: no cover
     """
     Removes the throughput file.
     """
-    os.remove(os.path.join(tempfile.gettempdir(), 'throughput'))
-    return True
+    if os.path.exists(statefile):
+        os.remove(os.path.join(tempfile.gettempdir(), 'throughput'))
+
 
 class Throughput(np.Resource):
-    statefile = os.path.join(tempfile.gettempdir(), 'throughput')
-
+    """
+    A throughput resource.
+    """
     def __init__(self, host, token, interface_name):
         self.host = host
         self.token = token
@@ -57,9 +63,9 @@ class Throughput(np.Resource):
 
     def probe(self):
         """
-        Querys the REST-API and create thermal metrics.
+        Querys the REST-API and create throughput metrics.
 
-        :return: a disk space metric.
+        :return: a throughput metric.
         """
 
         global api_inbytes, api_outbytes
@@ -73,8 +79,8 @@ class Throughput(np.Resource):
             api_inbytes = Finder.find_item(item, 'ibytes')
             api_outbytes = Finder.find_item(item, 'obytes')
 
-        _log.debug('Path to statefile: %r' % self.statefile)
-        with np.Cookie(self.statefile) as cookie:
+        _log.debug('Path to statefile: %r' % statefile)
+        with np.Cookie(statefile) as cookie:
             old_inbytes = cookie.get(self.interface_name + 'i', api_inbytes)
             old_outbytes = cookie.get(self.interface_name + 'o', api_outbytes)
             old_time = cookie.get(self.interface_name + 't', current_time)
@@ -114,6 +120,13 @@ class Throughput(np.Resource):
 
 
 class NetworkSummary(np.Summary):
+    """
+    Creates a throughput summary.
+    """
+
+    def __init__(self):
+        pass
+
     def ok(self, results):
         bit_in, bit_out = 0, 0
         for result in results:
@@ -122,10 +135,6 @@ class NetworkSummary(np.Summary):
             else:
                 bit_out += result.metric.value
         return 'Input is %s Mb/s - Output is %s Mb/s' % (
-            str(to_mega(bit_in)), str(to_mega(bit_out)))
+            str(Utils.to_mega(bit_in)), str(Utils.to_mega(bit_out)))
 
 
-def to_mega(size):
-    mega = 1000 * 1000
-    new_size = round(size / mega, 2)
-    return new_size
